add dp controller signal
